Pull TraversalUtility functions out of Cesium3DTilesetTraversal
